requirements.txt
# Keep Manager - Python Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Google API
google-auth==2.35.0
google-api-python-client==2.149.0

# Environment Variables
python-dotenv==1.0.1

# Data Validation (included with FastAPI but explicit for clarity)
pydantic==2.9.2
